test: mocks/ 아래 data/와 handlers/ 폴더를 추가하고 파일을 도메인 별로 관리
