+++ apps/web/src/mocks/data/questions.ts
@@ -1,0 +1,41 @@
+export const mockQuestions = {
+  questions: [
+    {
+      id: 1,
+      title: "TCP와 UDP의 차이점을 설명하세요",
+      content: "",
+      ttsUrl: null,
+      avgScore: 85,
+      avgImportance: 4.5,
+      categoryId: 11,
+      category: { id: 11, name: "TCP/IP", depth: 1, parentId: 1 },
+      score: null,
+    },
+    {
+      id: 2,
+      title: "HTTP와 HTTPS의 차이점은 무엇인가요?",
+      content: "",
+      ttsUrl: null,
+      avgScore: 78,
+      avgImportance: 4.0,
+      categoryId: 12,
+      category: { id: 12, name: "HTTP", depth: 1, parentId: 1 },
+      score: 90,
+    },
+    {
+      id: 3,
+      title: "프로세스와 스레드의 차이를 설명하세요",
+      content: "",
+      ttsUrl: null,
+      avgScore: 82,
+      avgImportance: 4.8,
+      categoryId: 21,
+      category: { id: 21, name: "프로세스", depth: 1, parentId: 2 },
+      score: null,
+    },
+  ],
+  totalCount: 3,
+  pageSize: 10,
+  currentPage: 1,
+  totalPages: 1,
+};
