+++ apps/web/src/mocks/handlers/index.ts
@@ -1,0 +1,4 @@
+import { categoryHandlers } from "./category-handlers";
+import { questionHandlers } from "./question-handlers";
+
+export const handlers = [...categoryHandlers, ...questionHandlers];
