+++ packages/types/tsconfig.tsbuildinfo
@@ -1,0 +1,1 @@
+{"fileNames":["../../node_modules/.pnpm/typescript@5.9.2/node_modules/typescript/lib/lib.d.ts","../../node_modules/.pnpm/typescript@5.9.2/node_modules/typescript/lib/lib.es5.d.ts","../../node_modules/.pnpm/typescript@5.9.2/node_modules/typescript/lib/lib.dom.d.ts","../../node_modules/.pnpm/typescript@5.9.2/node_modules/typescript/lib/lib.webworker.importscripts.d.ts","../../node_modules/.pnpm/typescript@5.9.2/node_modules/typescript/lib/lib.scripthost.d.ts","../../node_modules/.pnpm/typescript@5.9.2/node_modules/typescript/lib/lib.decorators.d.ts","../../node_modules/.pnpm/typescript@5.9.2/node_modules/typescript/lib/lib.decorators.legacy.d.ts","./src/domains/auth.ts","./src/domains/mypage.ts","./src/domains/questions.ts","./src/domains/reports.ts","./src/index.ts","../../node_modules/.pnpm/@types+estree@1.0.8/node_modules/@types/estree/index.d.ts","../../node_modules/.pnpm/@types+json-schema@7.0.15/node_modules/@types/json-schema/index.d.ts","../../node_modules/.pnpm/@types+eslint@9.6.1/node_modules/@types/eslint/use-at-your-own-risk.d.ts","../../node_modules/.pnpm/@types+eslint@9.6.1/node_modules/@types/eslint/index.d.ts","../../node_modules/.pnpm/@types+eslint-scope@3.7.7/node_modules/@types/eslint-scope/index.d.ts"],"fileIdsList":[[13,16],[13,14,15],[16]],"fileInfos":[{"version":"a7297ff837fcdf174a9524925966429eb8e5feecc2cc55cc06574e6b092c1eaa","impliedFormat":1},{"version":"c430d44666289dae81f30fa7b2edebf186ecc91a2d4c71266ea6ae76388792e1","affectsGlobalScope":true,"impliedFormat":1},{"version":"080941d9f9ff9307f7e27a83bcd888b7c8270716c39af943532438932ec1d0b9","affectsGlobalScope":true,"impliedFormat":1},{"version":"80e18897e5884b6723488d4f5652167e7bb5024f946743134ecc4aa4ee731f89","affectsGlobalScope":true,"impliedFormat":1},{"version":"cd034f499c6cdca722b60c04b5b1b78e058487a7085a8e0d6fb50809947ee573","affectsGlobalScope":true,"impliedFormat":1},{"version":"8e7f8264d0fb4c5339605a15daadb037bf238c10b654bb3eee14208f860a32ea","affectsGlobalScope":true,"impliedFormat":1},{"version":"782dec38049b92d4e85c1585fbea5474a219c6984a35b004963b00beb1aab538","affectsGlobalScope":true,"impliedFormat":1},"e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855","e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855","e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855","e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855","e917b35403c07f00132448a1cd8aaa1777210566a1bae1d654d4b001f908a7ff",{"version":"151ff381ef9ff8da2da9b9663ebf657eac35c4c9a19183420c05728f31a6761d","impliedFormat":1},{"version":"f3d8c757e148ad968f0d98697987db363070abada5f503da3c06aefd9d4248c1","impliedFormat":1},{"version":"a4a39b5714adfcadd3bbea6698ca2e942606d833bde62ad5fb6ec55f5e438ff8","impliedFormat":1},{"version":"bbc1d029093135d7d9bfa4b38cbf8761db505026cc458b5e9c8b74f4000e5e75","impliedFormat":1},{"version":"1f68ab0e055994eb337b67aa87d2a15e0200951e9664959b3866ee6f6b11a0fe","impliedFormat":1}],"root":[[8,12]],"options":{"composite":true,"declaration":true,"declarationMap":true,"outDir":"./dist","rootDir":"./src"},"referencedMap":[[17,1],[16,2],[15,3]],"semanticDiagnosticsPerFile":[[12,[{"start":14,"length":16,"messageText":"File '/Users/jihyokim/Projects/web11-MMH/packages/types/src/domains/auth.ts' is not a module.","category":1,"code":2306},{"start":46,"length":18,"messageText":"File '/Users/jihyokim/Projects/web11-MMH/packages/types/src/domains/mypage.ts' is not a module.","category":1,"code":2306},{"start":80,"length":21,"messageText":"File '/Users/jihyokim/Projects/web11-MMH/packages/types/src/domains/questions.ts' is not a module.","category":1,"code":2306},{"start":117,"length":19,"messageText":"File '/Users/jihyokim/Projects/web11-MMH/packages/types/src/domains/reports.ts' is not a module.","category":1,"code":2306}]],[15,[{"start":90,"length":3,"messageText":"Cannot find name 'Map'. Do you need to change your target library? Try changing the 'lib' compiler option to 'es2015' or later.","category":1,"code":2583},{"start":571,"length":16,"messageText":"Cannot find name 'IterableIterator'.","category":1,"code":2304}]],[16,[{"start":1567,"length":3,"messageText":"Cannot find name 'Map'. Do you need to change your target library? Try changing the 'lib' compiler option to 'es2015' or later.","category":1,"code":2583},{"start":36676,"length":16,"messageText":"Cannot find name 'IterableIterator'.","category":1,"code":2304},{"start":39333,"length":3,"messageText":"Cannot find name 'Map'. Do you need to change your target library? Try changing the 'lib' compiler option to 'es2015' or later.","category":1,"code":2583}]],[17,[{"start":755,"length":3,"messageText":"Cannot find name 'Map'. Do you need to change your target library? Try changing the 'lib' compiler option to 'es2015' or later.","category":1,"code":2583}]]],"latestChangedDtsFile":"./dist/index.d.ts","version":"5.9.2"}
